requirements.txt
matplotlib==3.10.1
numpy==2.2.5
pandas==2.2.3
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.2
timm==1.0.15
torch==2.6.0
torchgeo==0.7.0
torchmetrics==1.7.1
torchvision==0.21.0
tqdm==4.67.1
xarray==2024.3.0
